feat: add transient dashboard font override to GhosttySurfaceView
